drm: use u32 ids for LIST_LESSEES
Match the Linux KMS lease handler ABI for DRM_IOCTL_MODE_LIST_LESSEES: lessees_ptr is a 64-bit user pointer to an array of 32-bit lessee ids.

DragonFly had been writing uint64_t elements, which gives the wrong stride for Linux/libdrm-style callers even though the ioctl struct size is unchanged. Switch the handler and drmtest buffers to uint32_t and fix the local UAPI comment.

Add a sentinel stride gate so kms_smoke fails if the kernel writes past count_lessees with an 8-byte lessee-id stride.

Verification: drm.ko build; drmtest.c -Wall -Wextra -Werror; kms_smoke.py py_compile; git diff --check.

+++ sys/dev/drm/drm_lease.c
@@ -638,7 +638,7 @@ drm_mode_list_lessees_ioctl(struct drm_device *dev, void *data,
 	struct drm_mode_list_lessees *arg = data;
 	struct drm_master *owner;
 	struct drm_master *lessee;
-	uint64_t __user *lessee_ptr;
+	uint32_t __user *lessee_ptr;
 	uint32_t count = 0;
 
 	if (!drm_core_check_feature(dev, DRIVER_MODESET))
@@ -656,7 +656,7 @@ drm_mode_list_lessees_ioctl(struct drm_device *dev, void *data,
 		if (!drm_lease_master_has_objects_locked(lessee))
 			continue;
 		if (count < arg->count_lessees &&
-		    put_user((uint64_t)lessee->lessee_id, lessee_ptr + count)) {
+		    put_user((uint32_t)lessee->lessee_id, lessee_ptr + count)) {
 			mutex_unlock(&dev->mode_config.idr_mutex);
 			return -EFAULT;
 		}
